move listen events to separated file + (temporal I guess) file to store variables
