+++ Abominable-Intelligence/global_variables.py
@@ -1,0 +1,9 @@
+import os
+
+
+script_dir = os.path.dirname(os.path.abspath(__file__))
+config_path = os.path.join(script_dir, "config.yaml")
+
+# todo I guess: move to config and make function accessible only for owner to change these
+role_id_administration = 668245500612444205
+logs_channel = 732923500624347181
